feat(#1215): ICA: Shallow pass-through modules dilute locality
