+++ .pi/extensions/format-on-save/README.md
@@ -63,8 +63,7 @@ Part of Cheasee-Pi monorepo. Activated automatically.
 Adapter pattern with pluggable Formatter/Linter ports:
 
 ```
-├── index.ts               # Entry: registerHandler, tool_result event wiring
-├── formatting.mts         # looksLikeFilePath, MAX_FILE_SIZE_BYTES
+├── index.ts               # Entry: registerHandler, tool_result event wiring, looksLikeFilePath, MAX_FILE_SIZE_BYTES
 ├── eslint.mts             # formatEslintDiagnostics: diagnostic message formatting
 ├── eslint-adapter.mts     # EslintLinter: ESLint adapter (dynamic import)
 ├── prettier-adapter.mts   # PrettierFormatter: Prettier adapter (dynamic import)
